Improve the script that parsing the PASCAL annotations.
